requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Discord bot dependencies
discord.py>=2.3.0
aiohttp>=3.8.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0
